fix: image- null to old url & add: auctionId in auction command

+++ src/main/java/com/example/oisan/service/PostService.java
@@ -83,13 +83,17 @@ public class PostService {
     public Post updatePost(PostCommand postCom) {
         Customer customer = customerRepository.findCustomerByCustomerId(postCom.getCustomerId());
         String image_url = null;
-		try {
-			image_url = s3FileService.upload(postCom.getImage(), "post/");
-			s3FileService.deleteFile(image_url, "post/");
-			image_url = "post/"+image_url;
-		} catch (IOException e) {
-			e.printStackTrace();
-		}
+        if (postCom.getImage() == null) {
+        	image_url = postRepository.findByPostId(postCom.getPostId()).getImageUrl();
+        }else {
+			try {
+				image_url = s3FileService.upload(postCom.getImage(), "post/");
+				s3FileService.deleteFile(image_url, "post/");
+				image_url = "post/"+image_url;
+			} catch (IOException e) {
+				e.printStackTrace();
+			}
+        }
         Post post = new Post(postCom.getPostId(), customer, postCom.getCategId(), new Date(), postCom.getTitle(),
                 postCom.getDesc(), image_url, postCom.getWidth(), postCom.getHeight(), postCom.getDepth(), 1, postCom.getPrice());
         return postRepository.save(post);
